Refactor: Push state management up to the App component
