+++ editor/src/synths/AddPreset.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import FormElement from '../common/components/form/FormElement';
 import SynthSetting from './SynthSetting';
 import ChannelSetting from './ChannelSetting';
-import { PresetsAction } from './presetsReducer';
+import { PresetsAction } from '../common/state/settingsReducer';
 
 interface AddPresetProps {
   dispatch: React.Dispatch<PresetsAction>
